hardhat extension vs

+++ ethereum/contracts/BaseFund.sol
@@ -1,9 +1,7 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.16;
 
-
 contract BaseFundFactory {
-
   BaseFund[] public deployedBaseFunds;
 
   function getDeployedBaseFunds() public view returns (BaseFund[] memory) {
@@ -18,11 +16,17 @@ contract BaseFundFactory {
     bool _managersCanTransferMoneyWithoutARequest,
     bool _onlyManagersCanCreateARequest,
     bool _onlyContributorsCanApproveARequest,
-    uint _minimumContributionPercentageRequired,
-    uint _minimumApprovalsPercentageRequired
+    uint256 _minimumContributionPercentageRequired,
+    uint256 _minimumApprovalsPercentageRequired
   ) public {
-    require(_minimumContributionPercentageRequired < 101, 'Incorrect contribution percentage');
-    require(_minimumApprovalsPercentageRequired < 101, 'Incorrect approvals percentage');
+    require(
+      _minimumContributionPercentageRequired < 101,
+      "Incorrect contribution percentage"
+    );
+    require(
+      _minimumApprovalsPercentageRequired < 101,
+      "Incorrect approvals percentage"
+    );
 
     BaseFund newBaseFund = new BaseFund(
       _name,
@@ -37,21 +41,18 @@ contract BaseFundFactory {
     );
     deployedBaseFunds.push(newBaseFund);
   }
-
 }
 
-
 contract BaseFund {
-
   struct Request {
     string description;
     address petitioner;
     address recipient;
-    uint valueToTransfer;
-    uint transferredValue;
+    uint256 valueToTransfer;
+    uint256 transferredValue;
     bool complete;
     mapping(address => bool) approvals;
-    uint approvalsCount;
+    uint256 approvalsCount;
   }
 
   string public name;
@@ -62,28 +63,26 @@ contract BaseFund {
   bool public newManagersCanBeAdded;
 
   address[] public contributors;
-  mapping(address => uint) public contributions;
-  uint public totalContributions;
+  mapping(address => uint256) public contributions;
+  uint256 public totalContributions;
 
   bool public managersCanTransferMoneyWithoutARequest;
 
   Request[] public requests;
   bool public onlyManagersCanCreateARequest;
   bool public onlyContributorsCanApproveARequest;
-  uint public minimumContributionPercentageRequired;
-  uint public minimumApprovalsPercentageRequired;
-
+  uint256 public minimumContributionPercentageRequired;
+  uint256 public minimumApprovalsPercentageRequired;
 
   modifier onlyManagers() {
-    require(isManager[msg.sender], 'Only managers can access');
+    require(isManager[msg.sender], "Only managers can access");
     _;
   }
 
   modifier notManagers() {
-    require(!isManager[msg.sender], 'Managers can not access');
+    require(!isManager[msg.sender], "Managers can not access");
     _;
   }
-
 
   constructor(
     string memory _name,
@@ -93,8 +92,8 @@ contract BaseFund {
     bool _managersCanTransferMoneyWithoutARequest,
     bool _onlyManagersCanCreateARequest,
     bool _onlyContributorsCanApproveARequest,
-    uint _minimumContributionPercentageRequired,
-    uint _minimumApprovalsPercentageRequired
+    uint256 _minimumContributionPercentageRequired,
+    uint256 _minimumApprovalsPercentageRequired
   ) {
     name = _name;
     description = _description;
@@ -108,12 +107,12 @@ contract BaseFund {
   }
 
   function addNewManagers(address[] memory _managers) public {
-    require(newManagersCanBeAdded, 'New managers can not be added');
+    require(newManagersCanBeAdded, "New managers can not be added");
 
     _addManagers(_managers);
   }
 
-  function managersCount() public view returns (uint) {
+  function managersCount() public view returns (uint256) {
     return managers.length;
   }
 
@@ -125,25 +124,36 @@ contract BaseFund {
     _contribute(_for);
   }
 
-  function contributorsCount() public view returns (uint) {
+  function contributorsCount() public view returns (uint256) {
     return contributors.length;
   }
 
-  function balance() public view returns (uint) {
+  function balance() public view returns (uint256) {
     return address(this).balance;
   }
 
-  function transfer(address _to, uint _value) public {
-    require(managersCanTransferMoneyWithoutARequest, 'Managers can not transfer money without a request');
-    require(isManager[msg.sender], 'Only managers can access');
+  function transfer(address _to, uint256 _value) public {
+    require(
+      managersCanTransferMoneyWithoutARequest,
+      "Managers can not transfer money without a request"
+    );
+    require(isManager[msg.sender], "Only managers can access");
 
     payable(_to).transfer(_value);
   }
 
-  function createRequest(string memory _description, address _recipient, uint _valueToTransfer) public {
+  function createRequest(
+    string memory _description,
+    address _recipient,
+    uint256 _valueToTransfer
+  ) public {
     bool _isManager = isManager[msg.sender];
 
-    require(!onlyManagersCanCreateARequest || (onlyManagersCanCreateARequest && _isManager), 'Only managers can create a request');
+    require(
+      !onlyManagersCanCreateARequest ||
+        (onlyManagersCanCreateARequest && _isManager),
+      "Only managers can create a request"
+    );
 
     Request storage newRequest = requests.push();
 
@@ -153,33 +163,53 @@ contract BaseFund {
     newRequest.valueToTransfer = _valueToTransfer;
   }
 
-  function requestsCount() public view returns (uint) {
+  function requestsCount() public view returns (uint256) {
     return requests.length;
   }
 
-  function approveRequest(uint _index) public {
+  function approveRequest(uint256 _index) public {
     Request storage request = requests[_index];
 
-    require(!request.complete, 'The request has already been completed');
-    require((contributions[msg.sender] / totalContributions) * 100 >= minimumContributionPercentageRequired || (!onlyContributorsCanApproveARequest && isManager[msg.sender]), 'You can not approve a request');
-    require(!request.approvals[msg.sender], 'You have already approved this request');
+    require(!request.complete, "The request has already been completed");
+    require(
+      (contributions[msg.sender] / totalContributions) * 100 >=
+        minimumContributionPercentageRequired ||
+        (!onlyContributorsCanApproveARequest && isManager[msg.sender]),
+      "You can not approve a request"
+    );
+    require(
+      !request.approvals[msg.sender],
+      "You have already approved this request"
+    );
 
     request.approvals[msg.sender] = true;
     request.approvalsCount++;
   }
 
-  function finalizeRequest(uint _index) public {
+  function finalizeRequest(uint256 _index) public {
     Request storage request = requests[_index];
 
-    require(request.petitioner == msg.sender, 'You are not the petitioner of the request');
-    require(!request.complete, 'The request has already been completed');
+    require(
+      request.petitioner == msg.sender,
+      "You are not the petitioner of the request"
+    );
+    require(!request.complete, "The request has already been completed");
     if (onlyContributorsCanApproveARequest) {
-      require((request.approvalsCount / contributorsCount()) * 100 >= minimumApprovalsPercentageRequired, 'The request has not been approved yet');
+      require(
+        (request.approvalsCount / contributorsCount()) * 100 >=
+          minimumApprovalsPercentageRequired,
+        "The request has not been approved yet"
+      );
     } else {
-      require((request.approvalsCount / (managersCount() + contributorsCount())) * 100 >= minimumApprovalsPercentageRequired, 'The request has not been approved yet');
+      require(
+        (request.approvalsCount / (managersCount() + contributorsCount())) *
+          100 >=
+          minimumApprovalsPercentageRequired,
+        "The request has not been approved yet"
+      );
     }
 
-    uint _valueToTransfer = request.valueToTransfer;
+    uint256 _valueToTransfer = request.valueToTransfer;
     if (_valueToTransfer > balance()) {
       _valueToTransfer = balance();
     }
@@ -190,7 +220,7 @@ contract BaseFund {
   }
 
   function _addManagers(address[] memory _managers) private {
-    for (uint i; i < _managers.length;) {
+    for (uint256 i; i < _managers.length; ) {
       if (!isManager[msg.sender]) {
         managers.push(_managers[i]);
         isManager[_managers[i]] = true;
@@ -203,13 +233,12 @@ contract BaseFund {
   }
 
   function _contribute(address _contributor) private {
-    require(msg.value > 0, 'The contribution must be greater than zero');
+    require(msg.value > 0, "The contribution must be greater than zero");
 
     if (contributions[_contributor] == 0) {
       contributors.push(_contributor);
     }
     contributions[_contributor] += msg.value;
     totalContributions += msg.value;
-  }  
-
+  }
 }
